requirements.txt
python-dotenv
supabase
selenium
undetected-chromedriver
beautifulsoup4
requests
